Create PCORIACPRecruitment_2019-07-12_1110.REDCap.xml
This new project contains a series of new fields to capture the different clinic/arm assignments at the time that are relevant for the study analysis 9gre, orange, and red 'dots').
also, includes fiels to capture original and current PCP information.
ans finally, some indicators to handle out of office replies and other scenarios where PCP do not reply to investigators and a more manual approach needs to be taken to ensure that we follow the proper protocol before contacting patients directly.
Additional workflow charts shall be added to the documentation to explain how to use these new fields